+++ core/src/test/java/bisq/core/network/p2p/FileDatabaseTest.java
@@ -51,11 +51,7 @@ import org.junit.Test;
  *
  * TODO these tests are bound to be changed once Bisq migrates to a real database backend
  */
-public class FileDatabaseTest {
-    // Test fixtures
-    static final AccountAgeWitness object1 = new AccountAgeWitness(new byte[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}, 1);
-    static final AccountAgeWitness object2 = new AccountAgeWitness(new byte[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2}, 2);
-    static final AccountAgeWitness object3 = new AccountAgeWitness(new byte[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3}, 3);
+public class FileDatabaseTest extends FileDatabaseTestUtils {
 
     /**
      * TEST CASE: check if test fixture databases are in place and correct
@@ -275,91 +271,5 @@ public class FileDatabaseTest {
         Assert.assertTrue(result.contains(object3));
         // - did the historical data data store grow?
         Assert.assertEquals(1, DUT.getMap().size() - result.size());
-    }
-
-    /////////////////////////////////////////////////////////////////////////////////
-    ///////////////////////////////////// Utils /////////////////////////////////////
-    /////////////////////////////////////////////////////////////////////////////////
-
-    List<File> files = new ArrayList<>();
-    static final File storageDir = new File("src/test/resources");
-
-    @After
-    public void cleanup() {
-        try {
-            boolean done = false;
-            while (!done) {
-                Thread.sleep(100);
-                Set<Thread> threads = Thread.getAllStackTraces().keySet();
-                done = threads.stream().noneMatch(thread -> thread.getName().startsWith("Save-file-task"));
-            }
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        for (File file : files) {
-            file.delete();
-        }
-
-        try {
-            File backupDir = new File(storageDir + "/backup");
-            if (backupDir.exists())
-                Files.walk(backupDir.toPath()).sorted(Comparator.reverseOrder()).map(Path::toFile).forEach(File::delete);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        Arrays.stream(storageDir.list((dir, name) -> name.startsWith("AccountAgeWitnessStore"))).forEach(s -> {
-            new File(storageDir + File.separator + s).delete();
-        });
-    }
-
-    public File createFile(boolean isResourceFile, String name) {
-        File tmp;
-        if (isResourceFile)
-            tmp = new File(ClassLoader.getSystemClassLoader().getResource("").getFile() + File.separator + name);
-        else
-            tmp = new File(storageDir + File.separator + name);
-
-        files.add(tmp);
-        return tmp;
-    }
-
-    private AppendOnlyDataStoreService loadDatabase() {
-        Storage<AccountAgeWitnessStore> storage = new Storage<>(storageDir, new CorePersistenceProtoResolver(null, null, null, null), null);
-        AccountAgeWitnessStorageService storageService = new AccountAgeWitnessStorageService(storageDir, storage);
-        final AppendOnlyDataStoreService protectedDataStoreService = new AppendOnlyDataStoreService();
-        protectedDataStoreService.addService(storageService);
-        protectedDataStoreService.readFromResources("_TEST");
-        return protectedDataStoreService;
-    }
-
-    private void createDatabase(File target,
-                                AccountAgeWitness... objects) throws IOException {
-
-        if (null == objects) {
-            return;
-        }
-
-        String filename = "";
-        filename += Arrays.asList(objects).contains(object1) ? "o1" : "";
-        filename += Arrays.asList(objects).contains(object2) ? "o2" : "";
-        filename += Arrays.asList(objects).contains(object3) ? "o3" : "";
-
-        File source = new File(storageDir + File.separator + filename);
-
-        if (target.exists())
-            target.delete();
-
-        Files.copy(source.toPath(), target.toPath());
-    }
-
-    /**
-     * note that this function assumes a Bisq version format of x.y.z. It will not work with formats other than that eg. x.yy.z
-     * @param offset
-     * @return relative version string to the Version.VERSION constant
-     */
-    public String getVersion(int offset) {
-        return new StringBuilder().append(Integer.valueOf(Version.VERSION.replace(".", "")) + offset).insert(2, ".").insert(1, ".").toString();
     }
 }
